New new.schedb for 2016-17 school year.
This file was created from data scraped from WPI's new course scheduler.
The data is accurate as of Monday, 2016-04-18 (Monday; yesterday). The date at the top of the file may not be accurate.

Because the new scheduler includes data that this one is not meant to, there are a few differences to note:

1. Category II courses not offered in the year the file covers are not listed.

2. In many cases, sections are now separate from courses. This file lists every valid combination of subsections (lecture, lab, etc.) as a separate section.
For example, lecture A01 and lab AX02 will be listed as "A01 X02".

Each such section has the seats, waitlist, and CRN of its smallest subsection. Additionally, the CRN of each subsection is listed after the location.

Additionally, each "main" subsection (lectures, generally) is listed separately from those, with its own seats and waitlist visible.

For example, take a course with the following subsections:
Lecture A01: 100